feat: divide container and item components
