+++ src/components/ui/headerContainer.tsx
@@ -1,0 +1,7 @@
+export default function Header() {
+  return (
+    <header className="sticky top-0 z-10 bg-primary text-white">
+      <div className="container flex items-center justify-between h-14 py-4 px-6"></div>
+    </header>
+  );
+}
